% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{bayes_control}
\alias{bayes_control}
\title{Options for Bayesian deepregression}
\usage{
bayes_control(
  posterior_fun = posterior_mean_field,
  prior_fun = prior_trainable,
  kl_weight = function() 1/n_obs
)
}
\arguments{
\item{posterior_fun}{function defining the posterior function for the variational
verison of the network}

\item{prior_fun}{function defining the prior function for the variational
verison of the network}

\item{kl_weight}{KL weights for variational networks}
}
\value{
Returns a list with options
}
\description{
Options for Bayesian deepregression
}
