% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{check_and_install}
\alias{check_and_install}
\title{Function to check python environment and install necessary packages}
\usage{
check_and_install(force = FALSE, engine)
}
\arguments{
\item{force}{if TRUE, forces the installations}

\item{engine}{character; check if tf(= tensorflow) or torch is available}
}
\value{
Function that checks if a Python environment is available
and contains TensorFlow. If not the recommended version is installed.
}
\description{
If you encounter problems with installing the required python modules
please make sure, that a correct python version is configured using
\code{py_discover_config} and change the python version if required.
Internally uses \code{keras::install_keras}.
}
