% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_helpers.R
\name{check_input_args_fit}
\alias{check_input_args_fit}
\title{Function to check if inputs are supported by corresponding fit function}
\usage{
check_input_args_fit(args, fit_fun)
}
\arguments{
\item{args}{list; list of arguments used in fit process}

\item{fit_fun}{used fit function (e.g. fit.keras.engine.training.Model)}
}
\value{
stop message if inputs are not supported
}
\description{
Function to check if inputs are supported by corresponding fit function
}
