% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers_torch.R
\name{choose_kernel_initializer_torch}
\alias{choose_kernel_initializer_torch}
\title{Function to choose a kernel initializer for a torch layer}
\usage{
choose_kernel_initializer_torch(kernel_initializer, value = NULL)
}
\arguments{
\item{kernel_initializer}{string; initializer}

\item{value}{numeric; value used for a constant initializer}
}
\value{
kernel initializer
}
\description{
Function to choose a kernel initializer for a torch layer
}
