% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep-ensembles.R
\name{coef.drEnsemble}
\alias{coef.drEnsemble}
\title{Method for extracting ensemble coefficient estimates}
\usage{
\method{coef}{drEnsemble}(object, which_param = 1, type = NULL, ...)
}
\arguments{
\item{object}{object of class \code{"drEnsemble"}}

\item{which_param}{integer, indicating for which distribution parameter
coefficients should be returned (default is first parameter)}

\item{type}{either NULL (all types of coefficients are returned),
"linear" for linear coefficients or "smooth" for coefficients of 
smooth terms}

\item{...}{further arguments supplied to \code{coef.deepregression}}
}
\value{
list of coefficient estimates of all ensemble members
}
\description{
Method for extracting ensemble coefficient estimates
}
