% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families_torch.R
\name{collect_distribution_parameters}
\alias{collect_distribution_parameters}
\title{Character-to-parameter collection function needed for mixture of same distribution (torch)}
\usage{
collect_distribution_parameters(family)
}
\arguments{
\item{family}{character defining the distribution}
}
\value{
a list of extractions for each supported distribution
}
\description{
Character-to-parameter collection function needed for mixture of same distribution (torch)
}
