% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfuns.R
\name{combine_penalties}
\alias{combine_penalties}
\title{Function to combine two penalties}
\usage{
combine_penalties(penalties, dims)
}
\arguments{
\item{penalties}{a list of penalties}

\item{dims}{dimensions of the parameters to penalize}
}
\value{
a TensorFlow penalty combining the two penalties
}
\description{
Function to combine two penalties
}
