% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{compile.deepregression}
\alias{compile.deepregression}
\title{Compile an uncompiled deepregression object}
\usage{
\method{compile}{deepregression}(object, add_loss = NULL, weights = NULL, ind_fun = function(x) x)
}
\arguments{
\item{object}{deepregression object}

\item{add_loss}{an additional penalty added to the loss function}

\item{weights}{scalar or vector; weights used in the log-likelihood}

\item{ind_fun}{see \code{?deepregression_init}}
}
\description{
Compile an uncompiled deepregression object
}
