% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subnetwork_init.R
\name{layer_add_identity}
\alias{layer_add_identity}
\alias{layer_concatenate_identity}
\title{Convenience layer function}
\usage{
layer_add_identity(inputs)

layer_concatenate_identity(inputs)
}
\arguments{
\item{inputs}{list of tensors}
}
\value{
tensor
}
\description{
Convenience layer function
}
\details{
convenience layers to work with list of inputs where \code{inputs}
can also have length one
}
