% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\name{create_family}
\alias{create_family}
\title{Function to create (custom) family}
\usage{
create_family(tfd_dist, trafo_list, output_dim = 1L)
}
\arguments{
\item{tfd_dist}{a tensorflow probability distribution}

\item{trafo_list}{list of transformations h for each parameter 
(e.g, \code{exp} for a variance parameter)}

\item{output_dim}{integer defining the size of the response}
}
\value{
a function that can be used by 
\code{tfp$layers$DistributionLambda} to create a new 
distribuional layer
}
\description{
Function to create (custom) family
}
