% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families_torch.R
\name{create_family_torch}
\alias{create_family_torch}
\title{Function to create (custom) family}
\usage{
create_family_torch(torch_dist, trafo_list, output_dim = 1L)
}
\arguments{
\item{torch_dist}{a torch probability distribution}

\item{trafo_list}{list of transformations h for each parameter 
(e.g, \code{exp} for a variance parameter)}

\item{output_dim}{integer defining the size of the response}
}
\value{
a function that can be used to train a 
distribution learning model in torch
}
\description{
Function to create (custom) family
}
