% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psplinelayer.R
\name{create_penalty}
\alias{create_penalty}
\title{Function to create mgcv-type penalty}
\usage{
create_penalty(evaluated_gam_term, df, controls, Z = NULL)
}
\arguments{
\item{evaluated_gam_term}{a list resulting from a smoothConstruct call}

\item{df}{integer; specified degrees-of-freedom for the gam term}

\item{controls}{list; further arguments defining the smooth}

\item{Z}{matrix; matrix for constraint(s)}
}
\value{
a list with penalty parameter and penalty matrix
}
\description{
Function to create mgcv-type penalty
}
