% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{cv}
\alias{cv}
\title{Generic cv function}
\usage{
cv(x, ...)
}
\arguments{
\item{x}{model to do cv on}

\item{...}{further arguments passed to the class-specific function}
}
\description{
Generic cv function
}
