% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepregression.R
\name{deepregression_init}
\alias{deepregression_init}
\title{Initializing Deep Distributional Regression Models}
\usage{
deepregression_init(n_obs, ncol_structured, ncol_deep, list_structured,
  list_deep, input_pwr = NULL, use_bias_in_structured = TRUE,
  nr_params = 2, lss = TRUE, train_together = FALSE,
  lambda_lasso = NULL, family, dist_fun = NULL, variational = TRUE,
  weights = NULL, learning_rate = 0.01, optimizer = optimizer_adam(lr
  = learning_rate), monitor_metric = list(),
  mean_field = posterior_mean_field, prior = prior_trainable,
  orthog_fun = orthog, orthogX = NULL, inject_after_layer = rep(-1,
  length(list_deep)), residual_projection = FALSE, kl_weight = 1/n_obs)
}
\arguments{
\item{n_obs}{number of observations}

\item{ncol_structured}{a vector of length #parameters
defining the number of variables used for each of the parameters.
If any of the parameters is not modelled using a structured part
the corresponding entry must be zero.}

\item{ncol_deep}{a vector of length #parameters
defining the number of variables used for each of the parameters.
If any of the parameters is not modelled using a deep part
the corresponding entry must be zero. If all parameters
are estimated by the same deep model, the first entry must be
non-zero while the others must be zero.}

\item{list_structured}{list of (non-linear) structured layers
(list length between 0 and number of parameters)}

\item{list_deep}{list of deep models to be used
(list length between 0 and number of parameters)}

\item{residual_projection}{logical; defines which type of
projection is used.
If \code{TRUE}, the deep part is estimated together with the structured part;
if \code{FALSE}, the deep part is fitted on the residuals of the structured part.}
}
\description{
Initializing Deep Distributional Regression Models
}
