% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deeptransformation.R
\name{deeptransformation}
\alias{deeptransformation}
\title{Fitting Semi-Structured Transformation Models}
\usage{
deeptransformation(
  y,
  data,
  list_of_formulas,
  list_of_deep_models,
  order_bsp = NULL,
  y_basis_fun = function(y) eval_bsp(y, order = order_bsp, supp = range(y)),
  y_basis_fun_prime = function(y) eval_bsp_prime(y, order = order_bsp, supp =
    range(y))/diff(range(y)),
  split_between_shift_and_theta = NULL,
  addconst_interaction = NULL,
  base_distribution = "normal",
  ...
)
}
\arguments{
\item{y}{response variable}

\item{data}{data.frame or named list with input features}

\item{list_of_formulas}{a list of formulas for the shift term (h_2),
the interaction term (h_1) and, if, specified a third entry to specify 
lags for SATs models. See \code{?deepregression} for more details how
formulas can be specified.}

\item{list_of_deep_models}{a named list of functions
specifying a keras model.}

\item{order_bsp}{NULL or integer; order of Bernstein polynomials; if not NULL,
a conditional transformation model (CTM) is fitted.}

\item{y_basis_fun, y_basis_fun_prime}{basis functions for y transformation for CTM case}

\item{split_between_shift_and_theta}{if \code{family == 'transformation_model'} and
\code{!is.null(train_together)}, \code{split_between_shift_and_theta} is supposed
to define how many of the last layer's hidden units are used for the shift
term and how many for the theta term (an integer vector of length 2).}

\item{addconst_interaction}{positive constant;
a constant added to the additive predictor of the interaction term.
If \code{NULL}, terms are left unchanged. If 0 and predictors have negative values in their
design matrix, the minimum value of all predictors is added to ensure positivity.
If > 0, the minimum value plus the \code{addconst_interaction} is added to each predictor
in the interaction term.}

\item{base_distribution}{tfd_distribution or string; the base distribution for 
transformation models, see \code{?deeptransformation\_init}.}

\item{...}{further arguments passed to the \code{deepregression\_init} function}
}
\description{
Fitting Semi-Structured Transformation Models
}
