% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepregression.R
\name{distfun_to_dist}
\alias{distfun_to_dist}
\title{Function to define output distribution based on dist_fun}
\usage{
distfun_to_dist(dist_fun, preds)
}
\arguments{
\item{dist_fun}{a distribution function as defined by \code{make_tfd_dist}}

\item{preds}{tensors with predictions}
}
\value{
a symbolic tfp distribution
}
\description{
Function to define output distribution based on dist_fun
}
