% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distregression.R
\name{dr_init}
\alias{dr_init}
\title{Initializing a Distributional Regression Model}
\usage{
dr_init(
  n_obs,
  ncol_structured,
  list_structured,
  nr_params = 2,
  family,
  dist_fun = NULL,
  weights = NULL,
  monitor_metric = list(),
  output_dim = 1,
  mixture_dist = FALSE,
  ind_fun = function(x) x,
  extend_output_dim = 0,
  offset = NULL,
  additional_penalty = NULL,
  fsbatch_options = fsbatch_control(),
  optimizer = tf$keras$optimizers$SGD()
)
}
\arguments{
\item{n_obs}{number of observations}

\item{ncol_structured}{a vector of length #parameters
defining the number of variables used for each of the parameters.
If any of the parameters is not modelled using a structured part
the corresponding entry must be zero.}

\item{list_structured}{list of (non-linear) structured layers
(list length between 0 and number of parameters)}

\item{nr_params}{number of distribution parameters}

\item{family}{family specifying the distribution that is modelled}

\item{dist_fun}{a custom distribution applied to the last layer,
see \code{\link{make_tfd_dist}} for more details on how to construct
a custom distribution function.}

\item{weights}{observation weights used in the likelihood}

\item{monitor_metric}{see \code{?deepregression}}

\item{output_dim}{dimension of the output (> 1 for multivariate outcomes)}

\item{mixture_dist}{see \code{?deepregression}}

\item{ind_fun}{see \code{?deepregression}}

\item{extend_output_dim}{see \code{?deepregression}}

\item{offset}{list of logicals corresponding to the paramters;
defines per parameter if an offset should be added to the predictor}

\item{additional_penalty}{to specify any additional penalty, provide a function
that takes the \code{model$trainable_weights} as input and applies the
additional penalty. In order to get the correct index for the trainable
weights, you can run the model once and check its structure.}

\item{fsbatch_options}{options for Fellner-Schall algorithm, see 
\code{?deepregression}}

\item{optimizer}{see \code{?deepregression}}
}
\description{
Initializing a Distributional Regression Model
}
