% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep-ensembles.R
\name{ensemble}
\alias{ensemble}
\title{Generic deep ensemble function}
\usage{
ensemble(x, ...)
}
\arguments{
\item{x}{model to ensemble}

\item{...}{further arguments passed to the class-specific function}
}
\description{
Generic deep ensemble function
}
