% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psplinelayer.R
\name{extract_pure_gam_part}
\alias{extract_pure_gam_part}
\title{Extract the smooth term from a deepregression term specification}
\usage{
extract_pure_gam_part(term, remove_other_options = TRUE)
}
\arguments{
\item{term}{term specified in a formula}

\item{remove_other_options}{logical; whether to remove other options
withing the smooth term}
}
\value{
pure gam part of term
}
\description{
Extract the smooth term from a deepregression term specification
}
