% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psplinelayer.R
\name{extract_sp_S}
\alias{extract_sp_S}
\title{Convenience function to extract penalty matrix and value}
\usage{
extract_sp_S(x)
}
\arguments{
\item{x}{evaluated smooth term object}
}
\description{
Convenience function to extract penalty matrix and value
}
