% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_helpers.R
\name{extractval}
\alias{extractval}
\title{Extract value in term name}
\usage{
extractval(term, name, default_for_missing = FALSE, default = NULL)
}
\arguments{
\item{term}{character representing a formula term}

\item{name}{character; the value to extract}

\item{default_for_missing}{logical; if TRUE, returns \code{default} if argument is missing}

\item{default}{value returned when missing}
}
\value{
the value used for \code{name}
}
\description{
Extract value in term name
}
\examples{
extractval("s(a, la = 2)", "la")

}
