% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_helpers.R
\name{extractvar}
\alias{extractvar}
\title{Extract variable from term}
\usage{
extractvar(term, allow_ia = FALSE)
}
\arguments{
\item{term}{term specified in formula}

\item{allow_ia}{logical; whether to allow interaction of terms
using the \code{:} notation}
}
\value{
variable as string
}
\description{
Extract variable from term
}
