% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\name{family_to_tfd}
\alias{family_to_tfd}
\title{Character-tfd mapping function}
\usage{
family_to_tfd(family)
}
\arguments{
\item{family}{character defining the distribution}
}
\value{
a tfp distribution
}
\description{
Character-tfd mapping function
}
