% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families_torch.R
\name{family_to_trochd}
\alias{family_to_trochd}
\title{Character-torch mapping function}
\usage{
family_to_trochd(family)
}
\arguments{
\item{family}{character defining the distribution}
}
\value{
a torch distribution
}
\description{
Character-torch mapping function
}
