% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{fit}
\alias{fit}
\title{Generic train function}
\usage{
fit(object, ...)
}
\arguments{
\item{object}{object to apply fit on}

\item{...}{further arguments passed to the class-specific function}
}
\description{
Generic train function
}
