% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep-ensembles.R
\name{fitted.drEnsemble}
\alias{fitted.drEnsemble}
\title{Method for extracting the fitted values of an ensemble}
\usage{
\method{fitted}{drEnsemble}(object, apply_fun = tfd_mean, ...)
}
\arguments{
\item{object}{a deepregression model}

\item{apply_fun}{function applied to fitted distribution,
per default \code{tfd_mean}}

\item{...}{arguments passed to the \code{predict} function}
}
\value{
list of fitted values for each ensemble member
}
\description{
Method for extracting the fitted values of an ensemble
}
