% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{form_control}
\alias{form_control}
\title{Options for formula parsing}
\usage{
form_control(precalculate_gamparts = TRUE, check_form = TRUE)
}
\arguments{
\item{precalculate_gamparts}{logical; if TRUE (default), additive parts are pre-calculated
and can later be used more efficiently. Set to FALSE only if no smooth effects are in the 
formula(s) and a formula is very large so that extracting all terms takes long or might fail}

\item{check_form}{logical; if TRUE (default), the formula is checked in \code{process_terms}}
}
\value{
Returns a list with options
}
\description{
Options for formula parsing
}
