% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_helpers.R
\name{extractval}
\alias{extractval}
\alias{extractvals}
\alias{extractlen}
\alias{form2text}
\title{Formula helpers}
\usage{
extractval(term, name, default_for_missing = FALSE, default = NULL)

extractvals(term, names)

extractlen(term, data)

form2text(form)
}
\arguments{
\item{term}{formula term}

\item{name}{character; the value to extract}

\item{default_for_missing}{logical; if TRUE, returns \code{default} if argument is missing}

\item{default}{value returned when missing}

\item{names}{character vector of names}

\item{data}{a data.frame or list}

\item{form}{formula that is converted to a character string}
}
\value{
the value used for \code{name}
}
\description{
Formula helpers

Extractval with multiple options
}
\examples{
extractval("s(a, la = 2)", "la")

}
