% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepregression.R
\name{from_dist_to_loss}
\alias{from_dist_to_loss}
\title{Function to transform a distritbution layer output into a loss function}
\usage{
from_dist_to_loss(
  family,
  ind_fun = function(x) tfd_independent(x),
  weights = NULL
)
}
\arguments{
\item{family}{see \code{?deepregression}}

\item{ind_fun}{function applied to the model output before calculating the
log-likelihood. Per default independence is assumed by applying \code{tfd_independent}.}

\item{weights}{sample weights}
}
\value{
loss function
}
\description{
Function to transform a distritbution layer output into a loss function
}
