% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepregression_torch.R
\name{from_dist_to_loss_torch}
\alias{from_dist_to_loss_torch}
\title{Function to transform a distribution layer output into a loss function}
\usage{
from_dist_to_loss_torch(family, weights = NULL)
}
\arguments{
\item{family}{see \code{?deepregression}}

\item{weights}{sample weights}
}
\value{
loss function
}
\description{
Function to transform a distribution layer output into a loss function
}
