% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepregression_torch.R
\name{from_distfun_to_dist_torch}
\alias{from_distfun_to_dist_torch}
\title{Function to define output distribution based on dist_fun}
\usage{
from_distfun_to_dist_torch(dist_fun, preds)
}
\arguments{
\item{dist_fun}{a distribution function as defined by \code{make_torch_dist}}

\item{preds}{tensors with predictions}
}
\value{
a symbolic torch distribution
}
\description{
Function to define output distribution based on dist_fun
}
