% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{fsbatch_control}
\alias{fsbatch_control}
\title{Options for Fellner-Schall algorithm}
\usage{
fsbatch_control(
  factor = 0.01,
  lr_scheduler = NULL,
  avg_over_past = FALSE,
  constantdiv = 0,
  constantinv = 0,
  constinv_scheduler = NULL
)
}
\arguments{
\item{factor}{a factor defining how much of the past is used for the Fellner-Schall
algorithm; defaults to 0.01.}

\item{lr_scheduler}{a scheduler adapting
\code{factor} in each step; defaults to \code{NULL}.}

\item{avg_over_past}{logical, whether the beta coefficients should be averaged 
over the past values to stabilize estimation; defaults to \code{FALSE}}

\item{constantdiv}{small positive constant to stabilize training
in small batch regimes; defaults to 0.0.}

\item{constantinv}{small positive constant to stabilize training
in small batch regimes; defaults to 0.0.}

\item{constinv_scheduler}{scheduler for \code{constantinv}; per default 
NULL which results in an exponential decay with rate 1}
}
\value{
Returns a list with options
}
\description{
Options for Fellner-Schall algorithm
}
