% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psplinelayer.R
\name{gam_plot_data}
\alias{gam_plot_data}
\title{used by gam_processor}
\usage{
gam_plot_data(pp, weights, grid_length = 40, pe_fun = pe_gen)
}
\arguments{
\item{pp}{processed term}

\item{weights}{layer weights}

\item{grid_length}{length for grid for evaluating basis}

\item{pe_fun}{function used to generate partial effects}
}
\description{
used by gam_processor
}
