% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep-ensembles.R
\name{get_ensemble_distribution}
\alias{get_ensemble_distribution}
\title{Obtain the conditional ensemble distribution}
\usage{
get_ensemble_distribution(object, data = NULL, topK = NULL, ...)
}
\arguments{
\item{object}{object of class \code{"drEnsemble"}}

\item{data}{data for which to return the fitted distribution}

\item{topK}{not implemented yet}

\item{...}{further arguments currently ignored}
}
\value{
\code{tfd_distribution} of the ensemble, i.e., a mixture of the
    ensemble member's predicted distributions conditional on \code{data}
}
\description{
Obtain the conditional ensemble distribution
}
