% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psplinelayer.R
\name{get_gam_part}
\alias{get_gam_part}
\title{Extract gam part from wrapped term}
\usage{
get_gam_part(term, specials = c("s", "te", "ti"), wrapper = "vc")
}
\description{
Extract gam part from wrapped term
}
