% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psplinelayer.R
\name{get_gamdata}
\alias{get_gamdata}
\title{Extract property of gamdata}
\usage{
get_gamdata(
  term,
  param_nr,
  gamdata,
  what = c("data_trafo", "predict_trafo", "input_dim", "partial_effect", "sp_and_S",
    "df")
)
}
\arguments{
\item{term}{term in formula}

\item{param_nr}{integer; number of the distribution parameter}

\item{gamdata}{list as returned by \code{precalc_gam}}

\item{what}{string specifying what to return}
}
\value{
property of the gamdata object as defined by \code{what}
}
\description{
Extract property of gamdata
}
