% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psplinelayer.R
\name{get_gamdata_reduced_nr}
\alias{get_gamdata_reduced_nr}
\title{Extract number in matching table of reduced gam term}
\usage{
get_gamdata_reduced_nr(term, param_nr, gamdata)
}
\arguments{
\item{term}{term in formula}

\item{param_nr}{integer; number of the distribution parameter}

\item{gamdata}{list as returned by \code{precalc_gam}}
}
\value{
integer with number of gam term in matching table
}
\description{
Extract number in matching table of reduced gam term
}
