% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_helpers.R
\name{get_help_forward_torch}
\alias{get_help_forward_torch}
\title{Helper function to calculate amount of layers
Needed when shared layers are used, because of layers have same names}
\usage{
get_help_forward_torch(list_pred_param)
}
\arguments{
\item{list_pred_param}{list; subnetworks}
}
\value{
layers
}
\description{
Helper function to calculate amount of layers
Needed when shared layers are used, because of layers have same names
}
