% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfuns.R
\name{get_layernr_trainable}
\alias{get_layernr_trainable}
\title{Function to return layer numbers with trainable weights}
\usage{
get_layernr_trainable(mod, logic = FALSE)
}
\arguments{
\item{mod}{deepregression model}

\item{logic}{logical; TRUE: return logical vector; FALSE (default) index}
}
\description{
Function to return layer numbers with trainable weights
}
