% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_helpers.R
\name{get_luz_dataset}
\alias{get_luz_dataset}
\title{Helper function to create an function that generates R6 instances of 
class dataset}
\usage{
get_luz_dataset(df_list, target = NULL, length = NULL, object)
}
\arguments{
\item{df_list}{list; data for the distribution learning model (data for every distributional parameter)}

\item{target}{vector; target value}

\item{length}{amount of inputs}

\item{object}{deepregression object}
}
\value{
R6 instances of class dataset
}
\description{
Helper function to create an function that generates R6 instances of 
class dataset
}
