% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfuns.R
\name{get_names_pfc}
\alias{get_names_pfc}
\title{Extract term names from the parsed formula content}
\usage{
get_names_pfc(pfc)
}
\arguments{
\item{pfc}{parsed formula content}
}
\value{
vector of term names
}
\description{
Extract term names from the parsed formula content
}
