% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodelayer.R
\name{get_node_term}
\alias{get_node_term}
\title{Extract variables from wrapped node term}
\usage{
get_node_term(term)
}
\arguments{
\item{term}{character; node model term}
}
\value{
reduced variable node model term
}
\description{
Extract variables from wrapped node term
}
