% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodelayer.R
\name{get_nodedata}
\alias{get_nodedata}
\title{Extract attributes/hyper-parameters of the node term}
\usage{
get_nodedata(term, what)
}
\arguments{
\item{term}{term in formula}

\item{what}{string specifying what to return}
}
\value{
property of the node specification as defined by \code{what}
}
\description{
Extract attributes/hyper-parameters of the node term
}
