% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_partial_effect}
\alias{get_partial_effect}
\title{Return partial effect of one smooth term}
\usage{
get_partial_effect(
  object,
  names = NULL,
  return_matrix = FALSE,
  which_param = 1,
  newdata = NULL,
  ...
)
}
\arguments{
\item{object}{deepregression object}

\item{names}{string; for partial match with smooth term}

\item{return_matrix}{logical; whether to return the design matrix or}

\item{which_param}{integer; which distribution parameter
the partial effect (\code{FALSE}, default)}

\item{newdata}{data.frame; new data (optional)}

\item{...}{arguments passed to \code{get_weight_by_name}}
}
\description{
Return partial effect of one smooth term
}
