% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_helpers.R
\name{get_processor_name}
\alias{get_processor_name}
\title{Extract processor name from term}
\usage{
get_processor_name(term)
}
\arguments{
\item{term}{term in formula}
}
\value{
processor name as string
}
\description{
Extract processor name from term
}
