% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_helpers.R
\name{get_special}
\alias{get_special}
\title{Extract terms defined by specials in formula}
\usage{
get_special(term, specials, simplify = FALSE)
}
\arguments{
\item{term}{term in formula}

\item{specials}{string(s); special name(s)}

\item{simplify}{logical; shortcut for returning only
the name of the special in term}
}
\value{
specials in formula
}
\description{
Extract terms defined by specials in formula
}
