% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfuns.R
\name{get_type_pfc}
\alias{get_type_pfc}
\title{Function to subset parsed formulas}
\usage{
get_type_pfc(pfc, type = NULL)
}
\arguments{
\item{pfc}{list of parsed formulas}

\item{type}{either NULL (all types of coefficients are returned),
"linear" for linear coefficients or "smooth" for coefficients of}
}
\description{
Function to subset parsed formulas
}
