% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_weight_by_name}
\alias{get_weight_by_name}
\title{Function to retrieve the weights of a structured layer}
\usage{
get_weight_by_name(mod, name, param_nr = 1, postfixes = "")
}
\arguments{
\item{mod}{fitted deepregression object}

\item{name}{name of partial effect}

\item{param_nr}{distribution parameter number}

\item{postfixes}{character (vector) appended to layer name}
}
\value{
weight matrix
}
\description{
Function to retrieve the weights of a structured layer
}
