% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfuns.R
\name{get_weight_by_opname}
\alias{get_weight_by_opname}
\title{Function to return weight given model and name}
\usage{
get_weight_by_opname(mod, name, partial_match = FALSE)
}
\arguments{
\item{mod}{deepregression model}

\item{name}{character}

\item{partial_match}{logical; whether to also check for a partial match}
}
\description{
Function to return weight given model and name
}
