% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers_torch.R
\name{simplyconnected_layer_torch}
\alias{simplyconnected_layer_torch}
\alias{tiblinlasso_layer_torch}
\alias{tib_layer_torch}
\alias{tibgroup_layer_torch}
\title{Hadamard-type layers torch}
\usage{
simplyconnected_layer_torch(
  la = la,
  multfac_initializer = torch_ones,
  input_shape
)

tiblinlasso_layer_torch(
  la,
  input_shape = 1,
  units = 1,
  kernel_initializer = "he_normal"
)

tib_layer_torch(units, la, input_shape, multfac_initializer = torch_ones)

tibgroup_layer_torch(
  units,
  group_idx = NULL,
  la = 0,
  input_shape,
  kernel_initializer = "torch_ones",
  multfac_initializer = "he_normal"
)
}
\arguments{
\item{la}{numeric; regularization value (> 0)}

\item{input_shape}{integer; number of input dimension}

\item{units}{integer; number of units}

\item{kernel_initializer}{initializer}

\item{group_idx}{list of group indices}

\item{multfac_initialier}{initializer for parameters}
}
\value{
nn_module

torch layer object

nn_module

nn_module
}
\description{
Hadamard-type layers torch
}
