% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psplinelayer.R
\name{handle_gam_term}
\alias{handle_gam_term}
\title{Function to define smoothness and call mgcv's smooth constructor}
\usage{
handle_gam_term(object, data, controls)
}
\arguments{
\item{object}{character defining the model term}

\item{data}{data.frame or list}

\item{controls}{controls for penalization}
}
\value{
constructed smooth term
}
\description{
Function to define smoothness and call mgcv's smooth constructor
}
