% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers_torch.R
\name{layer_dense_torch}
\alias{layer_dense_torch}
\title{Function to define a torch layer similar to a tf dense layer}
\usage{
layer_dense_torch(
  input_shape,
  units = 1L,
  name,
  trainable = TRUE,
  kernel_initializer = "glorot_uniform",
  use_bias = FALSE,
  kernel_regularizer = NULL,
  ...
)
}
\arguments{
\item{input_shape}{integer; number of input units}

\item{units}{integer; number of output units}

\item{name}{string; string defining the layer's name}

\item{trainable}{logical; whether layer is trainable}

\item{kernel_initializer}{initializer; for coefficients}

\item{use_bias}{logical; wether bias is used (default no)}

\item{kernel_regularizer}{regularizer; for coefficients}

\item{...}{arguments used in choose_kernel_initializer_torch}
}
\value{
torch layer
}
\description{
Function to define a torch layer similar to a tf dense layer
}
