% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layer_sparse_batch_normalization}
\alias{layer_sparse_batch_normalization}
\title{Sparse Batch Normalization layer}
\usage{
layer_sparse_batch_normalization(lam = NULL, ...)
}
\arguments{
\item{lam}{regularization strength}

\item{...}{arguments passed to TensorFlow layer}
}
\value{
layer object
}
\description{
Sparse Batch Normalization layer
}
