% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layer_sparse_conv_2d}
\alias{layer_sparse_conv_2d}
\title{Sparse 2D Convolutional layer}
\usage{
layer_sparse_conv_2d(filters, kernel_size, lam = NULL, ...)
}
\arguments{
\item{filters}{number of filters}

\item{kernel_size}{size of convolutional filter}

\item{lam}{regularization strength}
}
\value{
layer object
}
\description{
Sparse 2D Convolutional layer
}
