% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psplinelayer.R
\name{layer_spline}
\alias{layer_spline}
\title{Function to define spline as TensorFlow layer}
\usage{
layer_spline(
  units = 1L,
  P,
  name,
  trainable = TRUE,
  kernel_initializer = "glorot_uniform"
)
}
\arguments{
\item{units}{integer; number of output units}

\item{P}{matrix; penalty matrix}

\item{name}{string; string defining the layer's name}

\item{trainable}{logical; whether layer is trainable}

\item{kernel_initializer}{initializer; for basis coefficients}
}
\value{
TensorFlow layer
}
\description{
Function to define spline as TensorFlow layer
}
