% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psplinelayer_torch.R
\name{layer_spline_torch}
\alias{layer_spline_torch}
\title{Function to define spline as Torch layer}
\usage{
layer_spline_torch(
  P,
  units = 1L,
  name,
  trainable = TRUE,
  kernel_initializer = "glorot_uniform",
  ...
)
}
\arguments{
\item{P}{matrix; penalty matrix}

\item{units}{integer; number of output units}

\item{name}{string; string defining the layer's name}

\item{trainable}{logical; whether layer is trainable}

\item{kernel_initializer}{initializer; for basis coefficients}

\item{...}{value used for constant kernel initializer}
}
\value{
Torch spline layer
}
\description{
Function to define spline as Torch layer
}
