% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{log_score}
\alias{log_score}
\title{Function to return the log_score}
\usage{
log_score(
  x,
  data = NULL,
  this_y = NULL,
  ind_fun = function(x) tfd_independent(x),
  convert_fun = as.matrix,
  summary_fun = function(x) x
)
}
\arguments{
\item{x}{the fitted deepregression object}

\item{data}{an optional data set}

\item{this_y}{new y for optional data}

\item{ind_fun}{function indicating the dependency; per default (iid assumption)
\code{tfd_independent} is used.}

\item{convert_fun}{function that converts Tensor; per default \code{as.matrix}}

\item{summary_fun}{function summarizing the output; per default the identity}
}
\description{
Function to return the log_score
}
