% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handler.R
\name{loop_through_pfc_and_call_trafo}
\alias{loop_through_pfc_and_call_trafo}
\title{Function to loop through parsed formulas and apply data trafo}
\usage{
loop_through_pfc_and_call_trafo(pfc, newdata = NULL)
}
\arguments{
\item{pfc}{list of processor transformed formulas}

\item{newdata}{list in the same format as the original data}
}
\value{
list of matrices or arrays
}
\description{
Function to loop through parsed formulas and apply data trafo
}
