% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-validation.R
\name{make_folds}
\alias{make_folds}
\title{Generate folds for CV out of one hot encoded matrix}
\usage{
make_folds(mat, val_train = 0, val_test = 1)
}
\arguments{
\item{mat}{matrix with columns corresponding to folds
and entries corresponding to a one hot encoding}

\item{val_train}{the value corresponding to train, per default 0}

\item{val_test}{the value corresponding to test, per default 1}
}
\description{
Generate folds for CV out of one hot encoded matrix
}
\details{
\code{val_train} and \code{val_test} can both be a set of value
}
