% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{make_generator}
\alias{make_generator}
\title{#' Generator for placeholder object
#'
#' @param var variable to be transformed; each entry must be a path to an
#' existing file
#' @param read_method function that is used to load the object from the path
#' @param dim the dimension of the actual data; if NULL, \code{read_method} is
#' applied to the first entry of \code{var}.
#'
#' @export
#'
path_to_placeholder <- function(var, read_method, dim = NULL)
{
  
  # check if data exists
  if(!all(sapply(var, file.exists)))
    stop("Not all files exist in the path variable. Can't create placeholder.")
  
  # check dimension
  if(is.null(dim))
    dim <- dim(read_method(var[1]))
  
  # create placeholder
  class(var) <- c("placeholder", "character")
  attr(var, "dims") <- c(length(var),dim)
  attr(var, "read_method") <- read_method
  
  return(var)
  
}}
\usage{
make_generator(
  data_image,
  data_tab,
  batch_size,
  target_size,
  color_mode,
  x_col,
  shuffle = TRUE,
  seed = 42L
)
}
\arguments{
\item{batch_size}{integer}

\item{x_col}{name of image column}

\item{shuffle}{logical for shuffling data}

\item{seed}{}
}
\description{
#' Methods for placeholder class
#' 
#' @method dim placeholder
#' @export
#' @rdname methodPlaceholder
dim.placeholder <- function(x) attr(x, "dims")
}
\details{
#' @export
#' @rdname methodPlaceholder
`[.placeholder` <- function(x, i, ...) {
  path_to_placeholder(NextMethod(),  read_method = attr(x, "read_method"), dim = attr(x, "dim"))
}

creates a generator for training
}
