% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{make_generator_from_matrix}
\alias{make_generator_from_matrix}
\title{Make a DataGenerator from a data.frame or matrix}
\usage{
make_generator_from_matrix(
  x,
  y = NULL,
  generator = image_data_generator(),
  batch_size = 32L,
  shuffle = TRUE,
  seed = 1L
)
}
\arguments{
\item{x}{matrix;}

\item{y}{vector;}

\item{generator}{generator as e.g. obtained from `keras::image_data_generator`.
Used for consistent train-test splits.}

\item{batch_size}{integer}

\item{shuffle}{logical; Should data be shuffled?}

\item{seed}{integer; seed for shuffling data.}
}
\description{
Creates a Python Class that internally iterates over the data.
}
