% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subnetwork_init.R
\name{makeInputs}
\alias{makeInputs}
\title{Convenience layer function}
\usage{
makeInputs(pp, param_nr)
}
\arguments{
\item{pp}{processed predictors}

\item{param_nr}{integer for the parameter}
}
\value{
input tensors with appropriate names
}
\description{
Convenience layer function
}
