% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_helpers.R
\name{makelayername}
\alias{makelayername}
\title{Function that takes term and create layer name}
\usage{
makelayername(term, param_nr, truncate = 60)
}
\arguments{
\item{term}{term in formula}

\item{param_nr}{integer; defining number of the distribution's parameter}

\item{truncate}{integer; value from which on names are truncated}
}
\value{
name (string) for layer
}
\description{
Function that takes term and create layer name
}
