% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.deepregression}
\alias{plot.deepregression}
\alias{prepare_data}
\alias{predict.deepregression}
\alias{fitted.deepregression}
\alias{fit.deepregression}
\alias{coef.deepregression}
\alias{print.deepregression}
\title{Generic functions for deepregression models}
\usage{
\method{plot}{deepregression}(object, which = NULL, which_param = 1,
  plot = TRUE, ...)

prepare_data(object, data)

\method{predict}{deepregression}(object, newdata = NULL,
  apply_fun = tfd_mean, convert_fun = as.matrix, ...)

\method{fitted}{deepregression}(object, apply_fun = tfd_mean, ...)

\method{fit}{deepregression}(object, ...)

\method{coef}{deepregression}(object)

\method{print}{deepregression}(object)
}
\arguments{
\item{object}{deepregression object}

\item{which}{which effect to plot, default selects all.}

\item{which_param}{integer of length 1.
Corresponds to the distribution parameter for which the effects should be plotted.}

\item{...}{further arguments, passed to fit, plot or predict function}

\item{newdata}{new data for prediction, defaults to \code{NULL}}

\item{apply_fun}{function to apply to distribution, default is \code{tfd_mean}.}

\item{convert_fun}{function which converts the tensor outputted by predict function}
}
\description{
Generic functions for deepregression models
}
