% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepregression_torch.R
\name{model_torch}
\alias{model_torch}
\title{Function to initialize a nn_module
Forward functions works with a list. The entries of the list are the input of 
the subnetworks}
\usage{
model_torch(submodules_list)
}
\arguments{
\item{submodules_list}{list; subnetworks}
}
\value{
nn_module
}
\description{
Function to initialize a nn_module
Forward functions works with a list. The entries of the list are the input of 
the subnetworks
}
