% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\name{multinorm_maker}
\alias{multinorm_maker}
\title{Implementation of a multivariate normal distribution}
\usage{
multinorm_maker(
  dim = 2,
  with_cov = TRUE,
  trafos_scale = exp,
  add_const = 1e-08
)
}
\arguments{
\item{dim}{dimension of the multivariate normal distribution}

\item{with_cov}{logical; whether or not to have a full covariance}

\item{trafos_scale}{transformation function for the scale}

\item{add_const}{small positive constant to stabilize calculations}
}
\description{
Implementation of a multivariate normal distribution
}
