% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{multioptimizer}
\alias{multioptimizer}
\title{Function to define an optimizer combining multiple optimizers}
\usage{
multioptimizer(optimizers_and_layers)
}
\arguments{
\item{optimizers_and_layers}{a list if \code{tuple}s of optimizer
and respective layers}
}
\value{
an optimizer
}
\description{
Function to define an optimizer combining multiple optimizers
}
