% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handler.R
\name{na_omit_list}
\alias{na_omit_list}
\title{Function to exclude NA values}
\usage{
na_omit_list(datalist)
}
\arguments{
\item{datalist}{list of data as returned by \code{prepare_data} and 
\code{prepare_newdata}}
}
\value{
list with NA values excluded and locations of original
NA positions as attributes
}
\description{
Function to exclude NA values
}
