% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\name{names_families}
\alias{names_families}
\title{Returns the parameter names for a given family}
\usage{
names_families(family)
}
\arguments{
\item{family}{character specifying the family as defined by \code{deepregression}}
}
\value{
vector of parameter names
}
\description{
Returns the parameter names for a given family
}
