% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers_torch.R
\name{nn_init_no_grad_constant_deepreg}
\alias{nn_init_no_grad_constant_deepreg}
\title{custom nn_linear module to overwrite reset_parameters
# nn_init_constant works only if value is scalar; so warmstarts for gam does'not work}
\usage{
nn_init_no_grad_constant_deepreg(tensor, value)
}
\arguments{
\item{tensor}{scalar or vector}

\item{value}{value used for constant initialization}
}
\value{
tensor
}
\description{
custom nn_linear module to overwrite reset_parameters
# nn_init_constant works only if value is scalar; so warmstarts for gam does'not work
}
