% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthogonalization.R
\name{orthog_P}
\alias{orthog_P}
\title{Function to compute adjusted penalty when orthogonalizing}
\usage{
orthog_P(P, Z)
}
\arguments{
\item{P}{matrix; original penalty matrix}

\item{Z}{matrix; constraint matrix}
}
\value{
adjusted penalty matrix
}
\description{
Function to compute adjusted penalty when orthogonalizing
}
