% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthogonalization.R
\name{orthog_post_fitting}
\alias{orthog_post_fitting}
\title{Orthogonalize a Semi-Structured Model Post-hoc}
\usage{
orthog_post_fitting(mod, name_penult, param_nr = 1)
}
\arguments{
\item{mod}{deepregression model}

\item{name_penult}{character name of the penultimate layer 
of the deep part part}

\item{param_nr}{integer; number of the parameter to be returned}
}
\value{
a \code{deepregression} object with weights frozen and
deep part specified by \code{name_penult} orthogonalized
}
\description{
Orthogonalize a Semi-Structured Model Post-hoc
}
