% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthogonalization.R
\name{orthog_structured_smooths_Z}
\alias{orthog_structured_smooths_Z}
\title{Orthogonalize structured term by another matrix}
\usage{
orthog_structured_smooths_Z(S, L)
}
\arguments{
\item{S}{matrix; matrix to orthogonalize}

\item{L}{matrix; matrix which defines the projection
and its orthogonal complement, in which \code{S} is projected}
}
\value{
constraint matrix
}
\description{
Orthogonalize structured term by another matrix
}
