% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfuns.R
\name{plot.drCV}
\alias{plot.drCV}
\title{Plot CV results from deepregression}
\usage{
\method{plot}{drCV}(x, what = c("loss", "weight"), ...)
}
\arguments{
\item{x}{\code{drCV} object returned by \code{cv.deepregression}}

\item{what}{character indicating what to plot (currently supported 'loss'
or 'weights')}

\item{...}{further arguments passed to \code{matplot}}
}
\description{
Plot CV results from deepregression
}
