% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-validation.R
\name{plot_cv}
\alias{plot_cv}
\title{Plot CV results from deepregression}
\usage{
plot_cv(x, what = c("loss", "weight"), ...)
}
\arguments{
\item{x}{\code{drCV} object returned by \code{cv.deepregression}}

\item{what}{character indicating what to plot (currently supported 'loss'
or 'weights')}

\item{...}{further arguments passed to \code{matplot}}
}
\description{
Plot CV results from deepregression
}
