% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psplinelayer.R
\name{precalc_gam}
\alias{precalc_gam}
\title{Pre-calculate all gam parts from the list of formulas}
\usage{
precalc_gam(lof, data, controls)
}
\arguments{
\item{lof}{list of formulas}

\item{data}{the data list}

\item{controls}{controls from deepregression}
}
\value{
a list of length 2 with a matching table to
link every unique gam term to formula entries and the respective
data transformation functions
}
\description{
Pre-calculate all gam parts from the list of formulas
}
