% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psplinelayer.R
\name{predict_gam_handler}
\alias{predict_gam_handler}
\title{Handler for prediction with gam terms}
\usage{
predict_gam_handler(object, newdata)
}
\arguments{
\item{object}{sterm}

\item{newdata}{data.frame or list}
}
\description{
Handler for prediction with gam terms
}
