% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{predict_gen}
\alias{predict_gen}
\title{Generator function for deepregression objects}
\usage{
predict_gen(
  object,
  newdata = NULL,
  batch_size = NULL,
  apply_fun = tfd_mean,
  convert_fun = as.matrix,
  ret_dist = FALSE
)
}
\arguments{
\item{object}{deepregression model;}

\item{newdata}{data.frame or list; for (optional) new data}

\item{batch_size}{integer; \code{NULL} will use the default (20)}

\item{apply_fun}{see \code{?predict.deepregression}}

\item{convert_fun}{see \code{?predict.deepregression}}

\item{ret_dist}{logical; whether to return the whole distribution or
only the (mean) prediction}
}
\value{
matrix or list of distributions
}
\description{
Generator function for deepregression objects
}
