% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handler.R
\name{prepare_data}
\alias{prepare_data}
\title{Function to prepare data based on parsed formulas}
\usage{
prepare_data(pfc, na_handler = na_omit_list, gamdata = NULL, engine = "tf")
}
\arguments{
\item{pfc}{list of processor transformed formulas}

\item{na_handler}{function to deal with NAs}

\item{gamdata}{processor for gam part}

\item{engine}{the engine which is used to setup the NN (tf or torch)}
}
\value{
list of matrices or arrays
}
\description{
Function to prepare data based on parsed formulas
}
