% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handler_torch.R
\name{prepare_data_torch}
\alias{prepare_data_torch}
\title{Function to additionally prepare data for fit process (torch)}
\usage{
prepare_data_torch(pfc, input_x, target = NULL, object)
}
\arguments{
\item{pfc}{list of processor transformed formulas}

\item{input_x}{output of prepare_data()}

\item{target}{target values}

\item{object}{a deepregression object}
}
\value{
list of matrices or arrays for predict or a dataloader for fit process
}
\description{
Function to additionally prepare data for fit process (torch)
}
