% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handler_torch.R
\name{prepare_input_list_model}
\alias{prepare_input_list_model}
\title{Function to prepare input list for fit process, due to different approaches}
\usage{
prepare_input_list_model(
  input_x,
  input_y,
  object,
  epochs = 10,
  batch_size = 32,
  validation_split = 0,
  validation_data = NULL,
  callbacks = NULL,
  verbose,
  view_metrics,
  early_stopping
)
}
\arguments{
\item{input_x}{output of prepare_data()}

\item{input_y}{target}

\item{object}{a deepregression object}

\item{epochs}{integer, the number of epochs to fit the model}

\item{batch_size}{integer, the batch size used for mini-batch training}

\item{validation_split}{float in [0,1] defining the amount of data used for validation}

\item{validation_data}{optional specified validation data}

\item{callbacks}{a list of callbacks for fitting}

\item{verbose}{logical, whether to print losses during training.}

\item{view_metrics}{logical, whether to trigger the Viewer in RStudio / Browser.}

\item{early_stopping}{logical, whether early stopping should be user.}
}
\value{
list of arguments used in fit function
}
\description{
Function to prepare input list for fit process, due to different approaches
}
