% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handler.R
\name{prepare_newdata}
\alias{prepare_newdata}
\title{Function to prepare new data based on parsed formulas}
\usage{
prepare_newdata(
  pfc,
  newdata,
  na_handler = na_omit_list,
  gamdata = NULL,
  engine = "tf"
)
}
\arguments{
\item{pfc}{list of processor transformed formulas}

\item{newdata}{list in the same format as the original data}

\item{na_handler}{function to deal with NAs}

\item{gamdata}{processor for gam part}

\item{engine}{character; the engine which is used to setup the NN (tf or torch)}
}
\value{
list of matrices or arrays
}
\description{
Function to prepare new data based on parsed formulas
}
