% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families_torch.R
\name{prepare_torch_distr_mixdistr}
\alias{prepare_torch_distr_mixdistr}
\title{Prepares distributions for mixture process}
\usage{
prepare_torch_distr_mixdistr(object, dists)
}
\arguments{
\item{object}{object of class \code{"drEnsemble"}}

\item{dists}{fitted distributions}
}
\value{
distribution parameters used for mixture of same distribution
}
\description{
Prepares distributions for mixture process
}
