% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_processing.R
\name{process_terms}
\alias{process_terms}
\title{Control function to define the processor for terms in the formula}
\usage{
process_terms(
  form,
  data,
  controls,
  output_dim,
  param_nr,
  parsing_options,
  specials_to_oz = c(),
  automatic_oz_check = TRUE,
  identify_intercept = FALSE,
  ...
)
}
\arguments{
\item{form}{the formula to be processed}

\item{data}{the data for the terms in the formula}

\item{controls}{controls for gam terms}

\item{output_dim}{the output dimension of the response}

\item{param_nr}{integer; identifier for the distribution parameter}

\item{parsing_options}{options}

\item{specials_to_oz}{specials that should be automatically checked for}

\item{automatic_oz_check}{logical; whether to automatically check for DNNs to be orthogonalized}

\item{identify_intercept}{logical; whether to make the intercept automatically identifiable}

\item{...}{further processors}
}
\value{
returns a processor function
}
\description{
Control function to define the processor for terms in the formula
}
