% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{quantile}
\alias{quantile}
\title{Generic quantile function}
\usage{
quantile(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{further arguments passed to the class-specific function}
}
\description{
Generic quantile function
}
