% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{re_layer}
\alias{re_layer}
\alias{pen_layer}
\title{random effect layer}
\usage{
re_layer(units, ...)

pen_layer(units, P, ...)
}
\arguments{
\item{units}{integer; number of units}

\item{...}{arguments passed to TensorFlow layer}

\item{P}{penalty matrix}
}
\value{
layer object

layer object
}
\description{
random effect layer

trainable penalty layer
}
