% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep-ensembles.R
\name{reinit_weights.deepregression}
\alias{reinit_weights.deepregression}
\title{Method to re-initialize weights of a \code{"deepregression"} model}
\usage{
\method{reinit_weights}{deepregression}(object)
}
\arguments{
\item{object}{object of class \code{"deepregression"}}
}
\value{
invisible \code{NULL}
}
\description{
Method to re-initialize weights of a \code{"deepregression"} model
}
