% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfuns.R
\name{separate_define_relation}
\alias{separate_define_relation}
\title{Function to define orthogonalization connections in the formula}
\usage{
separate_define_relation(
  form,
  specials,
  specials_to_oz,
  automatic_oz_check = TRUE
)
}
\arguments{
\item{form}{a formula for one distribution parameter}
}
\value{
Returns a list of formula components with ids and 
assignments for orthogonalization
}
\description{
Function to define orthogonalization connections in the formula
}
