% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers_torch.R
\name{simplyconnected_layer_torch}
\alias{simplyconnected_layer_torch}
\title{Custom nn module used for lasso layer}
\usage{
simplyconnected_layer_torch(
  la = la,
  multfac_initializer = torch_ones,
  input_shape
)
}
\value{
nn_module
}
\description{
Custom nn module used for lasso layer
}
