% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{stddev}
\alias{stddev}
\title{Generic sd function}
\usage{
stddev(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{further arguments passed to the class-specific function}
}
\description{
Generic sd function
}
