% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-validation.R
\name{stop_iter_cv_result}
\alias{stop_iter_cv_result}
\title{Function to get the stoppting iteration from CV}
\usage{
stop_iter_cv_result(
  res,
  thisFUN = mean,
  loss = "validloss",
  whichFUN = which.min
)
}
\arguments{
\item{res}{result of cv call}

\item{thisFUN}{aggregating function applied over folds}

\item{loss}{which loss to use for decision}

\item{whichFUN}{which function to use for decision}
}
\description{
Function to get the stoppting iteration from CV
}
