% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfuns.R
\name{tf_repeat}
\alias{tf_repeat}
\title{TensorFlow repeat function which is not available for TF 2.0}
\usage{
tf_repeat(a, dim)
}
\arguments{
\item{a}{tensor}

\item{dim}{dimension for repeating}
}
\description{
TensorFlow repeat function which is not available for TF 2.0
}
