% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfuns.R
\name{tf_row_tensor}
\alias{tf_row_tensor}
\title{Row-wise tensor product using TensorFlow}
\usage{
tf_row_tensor(a, b)
}
\arguments{
\item{a, b}{tensor}
}
\description{
Row-wise tensor product using TensorFlow
}
