% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_helpers.R
\name{tf_split_multiple}
\alias{tf_split_multiple}
\title{Split tensor in multiple parts}
\usage{
tf_split_multiple(A, len)
}
\arguments{
\item{A}{tensor}

\item{len}{integer; defines the split lengths}
}
\value{
list of tensors
}
\description{
Split tensor in multiple parts
}
