% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_helpers.R
\name{tf_stride_cols}
\alias{tf_stride_cols}
\title{Function to index tensors columns}
\usage{
tf_stride_cols(A, start, end = NULL)
}
\arguments{
\item{A}{tensor}

\item{start}{first index}

\item{end}{last index (equals start index if NULL)}
}
\value{
sliced tensor
}
\description{
Function to index tensors columns
}
