% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\name{tfd_mvr}
\alias{tfd_mvr}
\title{Implementation of a distribution-like layer for Mean-Variance Regression}
\usage{
tfd_mvr(
  loc,
  scale,
  validate_args = FALSE,
  allow_nan_stats = TRUE,
  name = "MVR"
)
}
\arguments{
\item{predictors}{a two dimensional predictor for mean and variance}
}
\description{
Implementation of a distribution-like layer for Mean-Variance Regression
}
