% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\name{tfd_zinb}
\alias{tfd_zinb}
\title{Implementation of a zero-inflated negbinom distribution for TFP}
\usage{
tfd_zinb(mu, r, probs)
}
\arguments{
\item{mu, r}{parameter of the negbin_ls distribution}

\item{probs}{vector of probabilites of length 2 (probability for poisson and
probability for 0s)}
}
\description{
Implementation of a zero-inflated negbinom distribution for TFP
}
