% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\name{tfd_zip}
\alias{tfd_zip}
\title{Implementation of a zero-inflated poisson distribution for TFP}
\usage{
tfd_zip(lambda, probs)
}
\arguments{
\item{lambda}{scalar value for rate of poisson distribution}

\item{probs}{vector of probabilites of length 2 (probability for poisson and
probability for 0s)}
}
\description{
Implementation of a zero-inflated poisson distribution for TFP
}
