% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers_torch.R
\name{tib_layer_torch}
\alias{tib_layer_torch}
\title{Custom nn_module for Hadamard-type layer}
\usage{
tib_layer_torch(units, la, input_shape, multfac_initializer = torch_ones)
}
\value{
nn_module
}
\description{
Custom nn_module for Hadamard-type layer
}
