% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers_torch.R
\name{tibgroup_layer_torch}
\alias{tibgroup_layer_torch}
\title{Custom nn_module for Group-Hadamard-type layer}
\usage{
tibgroup_layer_torch(
  units,
  group_idx = NULL,
  la = 0,
  input_shape,
  kernel_initializer = "torch_ones",
  multfac_initializer = "he_normal"
)
}
\value{
nn_module
}
\description{
Custom nn_module for Group-Hadamard-type layer
}
