% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\name{tweedie}
\alias{tweedie}
\title{tfd_distfun for (Quasi-)Tweedie to allow for flexible p}
\usage{
tweedie(p, quasi = FALSE, output_dim = 1L)
}
\arguments{
\item{p}{integer; defines distribution}

\item{quasi}{logical; whether to use quasi-likelihood or deviance resids}

\item{output_dim}{integer; currently only univariate responses supported}
}
\description{
tfd_distfun for (Quasi-)Tweedie to allow for flexible p
}
