% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{update_miniconda_deepregression}
\alias{update_miniconda_deepregression}
\title{Function to update miniconda and packages}
\usage{
update_miniconda_deepregression(
  python = VERSIONPY,
  uninstall = TRUE,
  also_packages = TRUE
)
}
\arguments{
\item{python}{string; version of python}

\item{uninstall}{logical; whether to uninstall previous conda env}

\item{also_packages}{logical; whether to install also all required packages}
}
\description{
Function to update miniconda and packages
}
