% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{weight_control}
\alias{weight_control}
\title{Options for weights of layers}
\usage{
weight_control(
  specific_weight_options = NULL,
  general_weight_options = list(activation = NULL, use_bias = FALSE, trainable = TRUE,
    kernel_initializer = "glorot_uniform", bias_initializer = "zeros", kernel_regularizer
    = NULL, bias_regularizer = NULL, activity_regularizer = NULL, kernel_constraint =
    NULL, bias_constraint = NULL),
  warmstart_weights = NULL,
  shared_layers = NULL
)
}
\arguments{
\item{specific_weight_options}{specific options for certain
weight terms; must be a list of length \code{length(list_of_formulas)} and
each element in turn a named list (names are term names as in the formula)
with specific options in a list}

\item{general_weight_options}{default options for layers}

\item{warmstart_weights}{While all keras layer options are availabe,
the user can further specify a list for each distribution parameter
with list elements corresponding to term names with values as vectors
corresponding to start weights of the respective weights}

\item{shared_layers}{list for each distribution parameter;
each list item can be again a list of character vectors specifying terms
which share layers}

\item{len}{integer; the length of \code{list_of_formulas}}
}
\value{
Returns a list with options
}
\description{
Options for weights of layers
}
