% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_and_install.R
\name{check_and_install}
\alias{check_and_install}
\title{Function to check python environment and install necessary packages}
\usage{
check_and_install(force = FALSE)
}
\arguments{
\item{force}{if TRUE, forces the installations}
}
\value{
Function that checks if a Python environment is available
and contains TensorFlow. If not the recommended version is installed.
}
\description{
Function to check python environment and install necessary packages
}
