% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{check_and_install}
\alias{check_and_install}
\title{Function to check python environment and install necessary packages}
\usage{
check_and_install(force = FALSE)
}
\arguments{
\item{force}{if TRUE, forces the installations}
}
\value{
Function that checks if a Python environment is available
and contains TensorFlow. If not the recommended version is installed.
}
\description{
If you encounter problems with installing the required python modules
please make sure, that a correct python version is configured using
`py_discover_config` and change the python version if required.
Internally uses keras::install_keras.
}
