% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperfuns.R
\name{extractval}
\alias{extractval}
\title{Extract value in term name}
\usage{
extractval(term, name, null_for_missing = FALSE)
}
\arguments{
\item{term}{character representing a formula term}

\item{name}{character; the value to extract}

\item{null_for_missing}{logical; if TRUE, returns NULL if argument is missing}
}
\value{
the value used for \code{name}
}
\description{
Extract value in term name
}
\examples{
extractval("s(a, la = 2)", "la")

}
