% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepregression.R
\name{from_preds_to_dist}
\alias{from_preds_to_dist}
\title{Define a Deep Distributional Regression Model}
\usage{
from_preds_to_dist(list_pred_param, family)
}
\arguments{
\item{list_pred_param}{list of input-output(-lists) generated from
\code{subnetwork_init}}

\item{family}{see \code{?deepregression}}
}
\value{
a list with input tensors and output tensors that can be passed
to, e.g., \code{keras_model}
}
\description{
Define a Deep Distributional Regression Model
}
