% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_distribution}
\alias{get_distribution}
\title{Function to return the fitted distribution}
\usage{
get_distribution(x, data = NULL, force_float = FALSE)
}
\arguments{
\item{x}{the fitted deepregression object}

\item{data}{an optional data set}

\item{force_float}{forces conversion into float tensors}
}
\description{
Function to return the fitted distribution
}
