% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psplinelayer.R
\name{get_gam_part}
\alias{get_gam_part}
\title{Extract gam part from wrapped term}
\usage{
get_gam_part(term, wrapper = "vc")
}
\arguments{
\item{term}{character; gam model term}

\item{wrapper}{character; function name that is wrapped around the gam part}
}
\description{
Extract gam part from wrapped term
}
