% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodelayer.R
\name{get_node_term}
\alias{get_node_term}
\title{Extract node part from wrapped term}
\usage{
get_node_term(term)
}
\arguments{
\item{term}{character; node model term}
}
\value{
reduced node model term
}
\description{
Extract node part from wrapped term
}
