% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepregression.R
\name{keras_dr}
\alias{keras_dr}
\title{Compile a Deep Distributional Regression Model}
\usage{
keras_dr(
  list_pred_param,
  family,
  weights = NULL,
  ind_fun = function(x) tfd_independent(x),
  optimizer = tf$keras$optimizers$Adam(),
  model_fun = keras_model,
  monitor_metrics = list(),
  ...
)
}
\arguments{
\item{list_pred_param}{list of input-output(-lists) generated from
\code{subnetwork_init}}

\item{family}{see \code{?deepregression}}

\item{weights}{vector of positive values; optional (default = 1 for all observations)}

\item{ind_fun}{function applied to the model output before calculating the
log-likelihood. Per default independence is assumed by applying \code{tfd_independent}.}

\item{optimizer}{optimizer used. Per default Adam}

\item{monitor_metrics}{Further metrics to monitor}

\item{model_instance}{which class of model to use (default \code{keras_model})}

\item{arguments}{passed to other functions}
}
\value{
a list with input tensors and output tensors that can be passed
to, e.g., \code{keras_model}
}
\description{
Compile a Deep Distributional Regression Model
}
