% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psplinelayer.R
\name{layer_spline}
\alias{layer_spline}
\title{Wrapper function to create spline layer}
\usage{
layer_spline(object, name = NULL, trainable = TRUE, input_shape,
  shared_weights = FALSE, lambdas, Ps, use_bias = FALSE,
  kernel_initializer = "glorot_uniform", bias_initializer = "zeros",
  ...)
}
\arguments{
\item{name}{An optional name string for the layer (should be unique).}

\item{input_shape}{Input dimensionality not including samples axis.}

\item{lambdas}{vector of smoothing parameters}

\item{Ps}{list of penalty matrices}

\item{Model}{or layer object.}
}
\description{
Wrapper function to create spline layer
}
