% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psplinelayer.R
\name{layer_spline}
\alias{layer_spline}
\title{Wrapper function to create spline layer}
\usage{
layer_spline(units = 1L, P, name)
}
\arguments{
\item{P}{a penalty matrix}

\item{name}{An optional name string for the layer (should be unique).}

\item{object}{or layer object.}

\item{trainable}{logical, whether the layer is trainable or not.}

\item{input_shape}{Input dimensionality not including samples axis.}

\item{kernel_initializer}{function to initialize the kernel (weight). Default
is "glorot_uniform".}

\item{bias_initializer}{function to initialize the bias. Default is 0.}

\item{variational}{logical, if TRUE, priors corresponding to the penalties
and posteriors as defined in \code{posterior_fun} are created}

\item{diffuse_scale}{diffuse scale prior for scalar weights}

\item{posterior_fun}{function defining the variational posterior}

\item{output_dim}{the number of units for the layer}

\item{...}{further arguments passed to \code{args} used in \code{create_layer}}
}
\description{
Wrapper function to create spline layer
}
