% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{make_generator}
\alias{make_generator}
\title{creates a generator for training}
\usage{
make_generator(
  data_image,
  data_tab,
  batch_size,
  target_size,
  color_mode,
  is_trafo = FALSE,
  x_col,
  shuffle = TRUE,
  seed = 42L
)
}
\arguments{
\item{data_image}{data.frame with image location as character}

\item{data_tab}{tabular data as given by \code{prepare_data}}

\item{batch_size}{integer}

\item{target_size}{size of the image without sample and colour channel}

\item{color_mode}{character; \code{'rgb'} or \code{'grayscale'}}

\item{is_trafo}{logical; indicating whether model is a trafo model}

\item{x_col}{name of image column in \code{data_image}}

\item{shuffle}{logical for shuffling data}

\item{seed}{seed for shuffling in generators}
}
\description{
creates a generator for training
}
