% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{make_generator}
\alias{make_generator}
\title{creates a generator for training}
\usage{
make_generator(
  input_x,
  input_y = NULL,
  batch_size,
  sizes,
  shuffle = TRUE,
  seed = 42L
)
}
\arguments{
\item{input_x}{list of matrices}

\item{input_y}{list of matrix}

\item{batch_size}{integer}

\item{sizes}{sizes of the image including colour channel}

\item{shuffle}{logical for shuffling data}

\item{seed}{seed for shuffling in generators}
}
\value{
generator for all x and y
}
\description{
creates a generator for training
}
