% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.deepregression}
\alias{plot.deepregression}
\alias{predict.deepregression}
\alias{fitted.deepregression}
\alias{fit.deepregression}
\alias{coef.deepregression}
\alias{print.deepregression}
\alias{cv}
\alias{mean.deepregression}
\alias{sd.deepregression}
\alias{quantile.deepregression}
\title{Generic functions for deepregression models}
\usage{
\method{plot}{deepregression}(x, which = NULL, which_param = 1, plot = TRUE, grid_length = 40, ...)

\method{predict}{deepregression}(
  object,
  newdata = NULL,
  batch_size = NULL,
  apply_fun = tfd_mean,
  convert_fun = as.matrix,
  dtype = "float32",
  ...
)

\method{fitted}{deepregression}(object, apply_fun = tfd_mean, ...)

\method{fit}{deepregression}(
  object,
  batch_size = NULL,
  epochs = 10,
  early_stopping = FALSE,
  verbose = TRUE,
  view_metrics = FALSE,
  patience = 20,
  save_weights = FALSE,
  validation_data = NULL,
  validation_split = 0.1,
  callbacks = list(),
  convertfun = function(x) tf$constant(x, dtype = "float32"),
  ...
)

\method{coef}{deepregression}(object, which_param = 1, type = NULL)

\method{print}{deepregression}(x, ...)

cv(
  x,
  verbose = FALSE,
  patience = 20,
  plot = TRUE,
  print_folds = TRUE,
  cv_folds = 5,
  stop_if_nan = TRUE,
  mylapply = lapply,
  save_weights = FALSE,
  callbacks = list(),
  ...
)

\method{mean}{deepregression}(x, data = NULL, ...)

\method{sd}{deepregression}(x, data = NULL, ...)

\method{quantile}{deepregression}(x, data = NULL, probs, ...)
}
\arguments{
\item{x}{a deepregression object}

\item{which_param}{integer, indicating for which distribution parameter
coefficients should be returned (default is first parameter)}

\item{plot}{whether to plot the resulting losses in each fold}

\item{grid_length}{the length of an equidistant grid at which a two-dimensional function
is evaluated for plotting.}

\item{...}{arguments passed to the \code{predict} function}

\item{object}{a deepregression model}

\item{newdata}{optional new data, either data.frame or list}

\item{batch_size}{}

\item{apply_fun}{function applied to fitted distribution,
per default \code{tfd_mean}}

\item{convert_fun}{how should the resulting tensor be converted,
per default \code{as.matrix}}

\item{dtype}{string for conversion}

\item{early_stopping}{logical, whether early stopping should be user.}

\item{verbose}{whether to print training in each fold}

\item{view_metrics}{logical, whether to trigger the Viewer in RStudio / Browser.}

\item{patience}{number of patience for early stopping}

\item{save_weights}{logical, whether to save weights in each epoch.}

\item{validation_data}{optional specified validation data}

\item{callbacks}{a list of callbacks used for fitting}

\item{convertfun}{function to convert R into Tensor object}

\item{type}{either NULL (all types of coefficients are returned),
"linear" for linear coefficients or "smooth" for coefficients of 
smooth terms}

\item{print_folds}{whether to print the current fold}

\item{cv_folds}{an integer if list with train and test data sets}

\item{stop_if_nan}{logical; whether to stop CV if NaN values occur}

\item{mylapply}{lapply function to be used; defaults to \code{lapply}}

\item{data}{either \code{NULL} or a new data set}

\item{probs}{the quantile value(s)}

\item{name}{which effect to plot as character; default plots all}

\item{auc_callback}{logical, whether to use a callback for AUC}
}
\value{
Returns an object \code{drCV}, a list, one list element for each fold
containing the model fit and the \code{weighthistory}.
}
\description{
Generic functions for deepregression models

Predict based on a deepregression object

Function to extract fitted distribution

Fit a deepregression model (pendant to fit for keras)

Extract layer weights / coefficients from model

Print function for deepregression model

Cross-validation for deepgression objects

mean of model fit

Standard deviation of fit distribution

Calculate the distribution quantiles
}
