% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\name{mix_dist_maker}
\alias{mix_dist_maker}
\title{generate mixture distribution of same family}
\usage{
mix_dist_maker(
  dist = tfd_normal,
  nr_comps = 3,
  trafos_each_param = list(function(x) x, function(x) tf$add(1e-08, tfe(x)))
)
}
\arguments{
\item{dist}{tfp distribution}

\item{nr_comps}{number of mixture components}

\item{trafos_each_param}{list of transformaiton applied before plugging
the linear predictor into the parameters of the distributions.
Should be of length #parameters of \code{dist}}
}
\value{
returns function than can be used as argument \code{dist\_fun} for
}
\description{
generate mixture distribution of same family
}
