% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handler.R
\name{prepare_data}
\alias{prepare_data}
\title{Function to prepare data based on parsed formulas}
\usage{
prepare_data(pfc)
}
\arguments{
\item{pfc}{list of processor transformed formulas}
}
\value{
list of matrices or arrays
}
\description{
Function to prepare data based on parsed formulas
}
