% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handler.R
\name{prepare_newdata}
\alias{prepare_newdata}
\title{Function to prepare new data based on parsed formulas}
\usage{
prepare_newdata(pfc, newdata, gamdata = NULL)
}
\arguments{
\item{pfc}{list of processor transformed formulas}

\item{newdata}{list in the same format as the original data}

\item{gamdata}{processor for gam part}
}
\value{
list of matrices or arrays
}
\description{
Function to prepare new data based on parsed formulas
}
