% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_processing.R
\name{processor}
\alias{processor}
\title{Control function to define the processor for terms in the formula}
\usage{
processor(
  form,
  data,
  controls,
  output_dim,
  param_nr,
  specials_to_oz = c(),
  automatic_oz_check = TRUE,
  ...
)
}
\arguments{
\item{data}{the data for the terms in the formula}

\item{controls}{controls for gam terms}

\item{param_nr}{integer; identifier for the distribution parameter}

\item{specials_to_oz}{specials that should be automatically checked for}

\item{automatic_oz_check}{logical; whether to automatically check for DNNs to be orthogonalized}

\item{...}{further processors}

\item{formula}{the formula to be processed}
}
\value{
returns a processor function
}
\description{
Control function to define the processor for terms in the formula
}
