% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep-ensembles.R
\name{reinit_weights}
\alias{reinit_weights}
\title{Genereic function to re-intialize model weights}
\usage{
reinit_weights(object, seed)
}
\arguments{
\item{object}{model to re-initialize}

\item{seed}{seed for reproducibility}
}
\description{
Genereic function to re-intialize model weights
}
