% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthogonalization.R
\name{separate_define_relation}
\alias{separate_define_relation}
\title{Function to define orthogonalization connections in the formula}
\usage{
separate_define_relation(
  form,
  specials,
  specials_to_oz,
  automatic_oz_check = TRUE,
  identify_intercept = FALSE,
  simplify = FALSE
)
}
\arguments{
\item{form}{a formula for one distribution parameter}

\item{specials}{specials in formula to handle separately}

\item{specials_to_oz}{parts of the formula to orthogonalize}

\item{automatic_oz_check}{logical; automatically check if terms must be orthogonalized}

\item{identify_intercept}{logical; whether to make the intercept identifiable}

\item{simplify}{logical; if FALSE, formulas are parsed more carefully.}
}
\value{
Returns a list of formula components with ids and 
assignments for orthogonalization
}
\description{
Function to define orthogonalization connections in the formula
}
