% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subnetwork_init.R
\name{subnetwork_init}
\alias{subnetwork_init}
\title{Initializes a Subnetwork based on the Processed Additive Predictor}
\usage{
subnetwork_init(
  pp,
  deep_top = NULL,
  orthog_fun = orthog_tf,
  split_fun = split_model,
  param_nr = 1
)
}
\arguments{
\item{pp}{processed predictor list from \code{processor}}

\item{deep_top}{keras layer if the top part of the deep network after orthogonalization
is different to the one extracted from the provided network}

\item{orthog_fun}{function used for orthogonalization}

\item{split_fun}{function to split the network to extract head}

\item{param_nr}{integer number for the distribution parameter}
}
\value{
returns a list of input and output for this additive predictor
}
\description{
Initializes a Subnetwork based on the Processed Additive Predictor
}
