% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\name{tfd_mse}
\alias{tfd_mse}
\title{For using mean squared error via TFP}
\usage{
tfd_mse(mean)
}
\arguments{
\item{mean}{parameter for the mean}
}
\description{
For using mean squared error via TFP
}
\details{
\code{deepregression} allows to train based on the
MSE by using \code{loss = "mse"} as argument to \code{deepregression}.
This tfd function just provides a dummy \code{family}
}
